[lib][debug] Make panic(...) invoke print once.
Previously panic would invoke print twice. Two calls give an opportunity
for the message to get fragmented in the log. Fusing the print calls is
an incremental step towards making writes to the log atomic.

Note this change relies on a GCC extension to deal with zero-length
vardic macros. All of the relevant compilers appear to support it,
however.

Bug: 122574827
Change-Id: I22890dcf6fadb55d756bfdf2885a519d908b44ed