MA-13885 add the SW workaround for i.MX8QM TKT340553
on i.MX8QM TO1.0, there is an issue: the bus width between A53-CCI-A72
is limited to 36bits.TLB maintenance through DVM messages over AR
channel,
some bits will be forced(truncated) to zero as the followings:

ASID[15:12] is forced to 0
VA[48:45] is forced to 0
VA[44:41] is forced to 0
VA[39:36] is forced to 0

This issue will result in the TLB aintenance across the clusters not
working
as expected due to some VA and ASID bits get truncated and forced to be
zero.

The SW workaround is: use the vmalle1is in Trusty OS due imx8 Trusty OS
will always use the VA that larger than 36bit.

Change-Id: I3ca33e0e836e1aa3363519460f1c63403b9bfb45
Signed-off-by: Haoran.Wang <elven.wang@nxp.com>