Handle cpu=-1 in get_top_thread
-1 is passed in for the cpu index to get_top_thread in order to find a
thread which could run on any CPU from the runqueue. If the runqueue is
empty, get_top_thread normally falls back to the idle thread for the
CPU it was asked about.

get_top_thread now returns NULL if the cpu it was passed is < 0, and the
caller which was passing in -1 handles a NULL return value.

Bug: 129083450
Change-Id: Id95fac7a4e2e014bc38a4b16a1f5010954506eba