[kernel][thread] Fix missed CPU wakeup IPIs
When we try to wake up a CPU to run a specific thread, we record the
priority of that thread so we don't request the same CPU again for a
lower priority thread. If another CPU picks up that thread before the
CPU we sent the wake-up IPI gets to the scheduler it could return
without updating the priority again as it has an early return path if no
context switch is needed.

Fix the early return path to also set the priority. Also fix a priority
update that was disabled when a debug flag was diabled.

Bug: 129082895
Change-Id: I37b7bcea82650bb2a017bdd7374ab037a3919e88