Use functions instead of macros for alignment
Switching ROUNDUP/ROUNDDOWN/ALIGN to static inline functions prevents
the types of these values from being implicit.

While I did not find any examples of it causing problems, the previous
approach could cause different behavior depending on the relative types
of the alignment and value to align.
For example, if an unsigned literal was used in the alignment field (say
8U) you would end up dropping the high 32 bits of the pointer, since no
sign extension would occur when extending the mask to the width of the
pointer.

This also prevents UBSan from alarming on the implicit casts that were
previously performed in the macro.

Change-Id: I43c94ecd0f6cce53d4a0638139f39a0e5533f407