Adjust list_for_every_entry macros to avoid UB
We were doing out-of-bounds address math in order to iterate over
entries. We fabricated a fake thread_t * out of our head list node,
which leads to invoking containerof on a value which is not necessarily
embedded in that container.

UBSan caught this because the resulting thread_t * was not properly
aligned.

Since we are holding a pointer to the next list node, we also remove
temp_entry from the API.

Bug: 129300035
Change-Id: I117e043dcfd3aa8001100d7dd97e71c87e56316c