Add TLS entry for UBSan state
We need to track this so that if a thread is already trying to report a
UBSan fault, it won't retrigger the fault handler code by trying to
report it.

This value is thread-specific because we are only trying to prevent two
UBSan handlers in the same callstack with this flag - in different
threads, both handlers should report as normal.

Bug: 129300035
Change-Id: If823e0ab7c3afbe72fa2d15b8e89d6b9cd6a9398