[lib][heap][miniheap] Remove retry limit after growing heap
If multiple threads tries to allocate a page at the same time, retrying
the allocation can fail as the heap lock is released between growing the
heap and retrying the allocation. Remove the retry limit so that it will
succeed if memory is still available after the other threads are done.

Bug: 132652075
Change-Id: Ic62a773a49288db1f15fc52abaeaef6d7a94d770