vmm: Create vmm_obj
Create an object type that can be mapped by the vmm. This allows
tracking when memory is no longer mapped so that it can be freed or
returned. Currently the vmm keeps track of a pages that need to be freed
when a vmm region is unmapped, but this does not work for shared memory
where the physical memory could be mapped multiple times, or where the
physical memory was not allocated by lk.

Bug: 129788581
Change-Id: I200f7ac10448ee54b15d18784ff76011d76777df