[arch][arm] Fix fiq handler for ARM_MERGE_FIQ_IRQ
When entering most exceptions on arm, IRQs are masked but FIQs are left
unmasked. Simulate the behavior we want by looking at the IRQ mask in
SPSR on FIQ entry and restart with FIQs masked as well if only IRQs were
masked. This prevents corrupting the LR and SPSR values for the
interrupted context as FIQs can trigger a context switch when
ARM_MERGE_FIQ_IRQ is enabled.

Bug: 122357256
Change-Id: Ice0f5bee13bc4a52eba3a82ce945c9c693d9b224