[kernel][timer] Fix crash caused by timer_cancel.
If timer_cancel is called while the timer callback is about to run on
another cpu, the timer interrupt will crash jumping to the now NULL
callback pointer. If the timer is freed after timer_cancel returns
similar crashes are also possible.

Add timer_cancel_sync, which waits for the timer callback to return, and
use it where the timer could be running on another cpu.

Bug: 145226667
Change-Id: Id3e43dfa54a9460440c7312af226e04e42953abc