arm_gic: Trigger SGIs with GICv3 affinity routing
GICv3 uses a different format for the SGIR register, defining writes to
it based on affinity values. This adds support for this type of routing,
and allows platform-specific override of the mapping from logical cpu
number to affinities via arch_cpu_num_to_gic_affinities in case an
architecture has a more exotic affinity layout. A platform will need to
define this if they are using GICv3 with affinity routing, and have a
custom arch_curr_cpu_num implementation.

Bug: 143179093
Change-Id: I7896cc88c910144d0df67c590ae926707c8afa55