[kernel] Add thread flag to exit thread on panic
If panic is called from a thread after
thread_set_flag_exit_on_panic(thread, true) has been called for that
thread. The thread will exit with ERR_FAULT as the return code instead
of halting the system. This can be used to test error paths that are
normally not testable, e.g. exception handling.

Bug: 147698516
Change-Id: I455898d8d3b5b3ee005969893bb4cb09d4b88a48