[kernel][thread] Add thread_sleep_until_ns
Add api to sleep until a specific time. If that time is in the past,
return immediately. The underlying timer api is still relative so there
is no improved precision over using the relative api, but by using
thread_sleep_until_ns the client does not have to compute the relative
time and deal with the pitfall of that computation undeflowing when the
target time is already in the past.

Bug: 162252601
Change-Id: Ieb1b6a790ba67a3f625de8fa03576bfcd590686f