[arch][arm/arm64] Handle stack overflows
Currently a stack overflow that triggers a fault when hitting a guard
page, will trigger an exception loop until the exception handler hits
another writable memory mapping. To prevent that check the stack pointer
before touching it (arm) or use a separate mode while pusing the iframe
so a fault there does not reenter the same code (arm64).

Bug: 165825378
Change-Id: I767fd7acc74717120a87d207cdf04655d813fb6a