[arch][arm64] Outline asm in arch_enter_uspace
Move inline assembly to arch/arm64/exceptions.S which contains
other exception return paths in preparation for kernel SCS. This
eliminates the need to clobber registers in the inline version to
avoid collision with the input registers in the inline version.

This is a non-functional change except that x14 is zeroed. This
makes the initial stack frame consistent between aarch64 and
arm32 w.r.t. the link register.

Test: build.py qemu-generic-arm64-test-debug
Bug: None
Change-Id: If770e57d4acf2b901fb420dc37b9ddb850ad4e32