[make] Force correct frame pointers for sancov
Accurate LLVM coverage instrumentation with trace-pc-guard requires
being able to correctly find a function's caller. With sibling call
optimization, calls at the end of a function are turned into a direct
branch to the sibling function, which causes the builtin returning our
caller in the instrumentation to return the wrong caller.

Disable sibling call optimization when building with coverage
instrumentation.

Bug: 175221942
Test: USER_COVERAGE_ENABLED=true build.py qemu-generic-arm64-fuzz-test-debug
Change-Id: I596ee3565a95d288f2633042ded9573ce0583915