[dev][interrupt][arm_gic] Add DSB after igrpen write
This register is not self-synchronising. Without the DSP,
sm_intc_enable_interrupts could return before the pending interrupt was
visible to the CPU and libsm would return to the non-secure world
without handling the interrupt. The FIQ would then trigger again and
trusty would in turn trigger a new doorbell interrupt resulting in a
loop.

Bug: 170482177
Change-Id: If63f6d0b1bf0631edb6fff22220b33665e45aec7