vmm: Put guard page at end of address space
By default, make the last page of an address space a guard page.
This prevents a pointer overflow, which is undefined behavior,
when an allocation is at the very end of the address space and
pointer arithmetic computes a pointer just beyond the end of
an allocation. A specific allocation request can still map the
last page of an address space if the VMM_FLAG_NO_END_GUARD flag
is set.

Bug: 129300035
Change-Id: Ie9480b0c7ebde220da90055f64e20da984edea6f