vmm: Guard non-specific allocations and remove unused args
This change mandates that non-specific (i.e., randomly
placed) allocations have guard pages. This means that the
address returned by vmm_find_spot will now include space
for guard pages above and below the gap if one was found.

The vmm_flags arguments passed to alloc_spot, extract_gap,
scan_gap, and spot_in_gap became redundant with this change
and were removed.

The alloc_region function now returns a status code rather
than a pointer to the allocated region (or a null pointer
in case of an error). The allocated region structure is now
passed via the out parameter if the operation succeeds. If
alloc_region or the add_region_to_aspace helper function
fails, the status code indicates the type of the error.

Change-Id: Idb06a81be8d24f5bec4cd7038afe3631b61d7939