[x86] Adjust user address space and export as a define
As of https://android-review.googlesource.com/c/trusty/lib/+/1483358 we
use USER_ASPACE_BASE in the Trusty libc and assume that it is not zero.
This did not cause an issue initially because we were not building libc
for x86 Trusty at the time.

This change moves the x86 address space base forward one page and
exports this constant (along with the address space size) as a
preprocessor define.

Test: build.py generic-x86_64
Bug: None
Change-Id: I8a975a8701ec2ff82312657af0165754e00174b2