[arch][arm64] Prevent kernel and user from executing each other's code
Set UXN on executable kernel pages and PXN on executable user
pages to prevent the kernel and user space from executing each
other's code. This is a port of Fuchsia change 608503.
This change is needed to preserve the current permissions
after porting the initial mapping code from assembly to C.

Bug: 80147716
Bug: 207152606
Change-Id: Iee9855e8c52e9358706b23935034a73f566110ba