MA-20249-2 imx8mq: Add workaround code for ERR11171
According to ERR11171, SGIs on imx8mq can't wakeup CPU
cores when it's in idle state. Workaround path has been
added in TF-A to wakeup the CPU cores with IRQ32 after
each IPIs are triggered.

This commit follows the same workaround path to wakeup
the cpu core with IRQ32 after each IPI.

Change-Id: I703bf0ef016ea1f2b4bbffd6120e76db376a4d26
Signed-off-by: Ji Luo <ji.luo@nxp.com>