[make] Fix defines containing quotes
MODULE_DEFINES gets emitted in module_config.h to ensure that changes to
compilation flags result in re-compilation of affected sources. If any
flags contained quotes, these quotes would result in preprocessor
defines with embedded quote characters, which happens to parse in C but
not in C++.

This change escapes quotes in flags for use in module_config.h so we
don't have to worry about this difference in parsing.

Test: build.py qemu-generic-arm64-test-debug and inspect module_config.h
Bug: 232019245
Change-Id: I59a1ac91d9161fb35e7e9c8649640f052eb3cf0a