lib/libc: Make vfprintf definition match musl
Musl's stdio.h defines `vfprintf` as an inline function that
calls `vfprintf_worker` with a fourth argument that
determines whether address-like values should be printed or
not. To use musl for the kernel, this commit renames the
`vfprintf` definition in stdio.c to `vfprintf_worker` and
changes the function signature to match musl.

Bug: 230134581
Change-Id: I94dd4100c06b0a23d55a0ccdbcd3e2dee4f40a10