[lib][libc] Improving log filtering capabilities
Adding modifiers to override filtering. We are using the same modifier
used by printk in the linux kernel to override filtering of pointers
(%px), and extending it to also %dx, %ux and %xx.
We are also removing the override of RELEASE_BUILD on code because
this is now taken care on test build makefiles.

Bug: 234061145
Test: Added host unit tests for both test-debug and release like builds.
Change-Id: I120cd7ef6664ec02ec02c2324cfa9bac3445e6a6