Switch to PRI*64 macros for format specifiers
LK defines uint64_t as `unsigned long long` for every arch,
but musl defines it as `unsigned long` for aarch64 and x64
so this switches %llx to PRIx64 when printing `uint64_t`s.
There are analogous changes for %llu and %lld. This also
adds a PRI macro for printing `asid_t`.

Bug: 230134581
Change-Id: Iafa6b4c1d341832e3aeead0ee3fc517b8e3b1ff6