lib/libc: Make DEBUG_ASSERT use asserted expression in all cases
When LK_DEBUGLEVEL is less than 2 the variable asserted in
DEBUG_ASSERT must be used afterwards to avoid a `variable
set but not used` error. This change allows defining
variables which only need to be asserted. It also fixes a
few cases in miniheap that did not compile because of
conditionally defined struct fields/variables. This fixes
the generic-arm* targets when switching to musl.

Bug: 230134581
Change-Id: I9753d238470caabaaebd5fb02558fc37ebce94ba