lib/libc: Add fd_io_handle and file_io_handle
LK's libc exposes the FILE struct's definition which
app/consoletest and libtrusty use to access the
`io_handle_t`s corresponding to stdout/stderr. On the other
hande, musl only exposes an opaque type for FILE. This
commit adds two new functions for accessing the
`io_handle_t`s given a file descriptor or a FILE*. These
functions also have equivalents for musl defined in
libc-trusty. This allows us to modify the modules that map
fd or FILE* to io_handle_t with a call to one of the new
functions and select the definition based on which libc
implementation is chosen.

Bug: 230134581
Change-Id: I2052258614b42413cc6ea040bd29b42249cc7c04