lib/libc: Add option to replace LK libc with musl
This commit adds an LK_LIBC_IMPLEMENTATION build flag that
can be set to `musl` or `lk` to select the kernel's libc.
The default (musl) is set in engine.mk to allow projects
that want to keep using lk to override it. This is also
required to allow conditionally building tests in
kerneltests-inc.mk as libc-trusty/test only compiles with
musl (LK libc is missing many headers and function
definitions). When a libc module is selected in
kernel/rules.mk we also define an
LK_LIBC_IMPLEMENTATION_IS_[LK,MUSL] macro to allow
conditionally compiling code based on which libc is used if
it's necessary.

This commit also moves off_t and ssize_t from lk/types.h to
LK's sys/types.h to only keep the LK definitions (which
differ from musl on some archs) when using LK libc. It also
includes inttypes.h in lk/types.h (used by both musl and
LK) to allow using the PRI macros it defines.

Bug: 230134581
Change-Id: I7a6c4b6d50d206241775e83961ee166689906515