top/main.c: Disable CFI check for ctors when building with C++
This commit disables a CFI check when calling ctors. This is
currently needed to call constructors inserted by clang for
global C++ objects since clang currently does not emit CFI
type info for these functions.  When we're not building
libcxx in the kernel `call_constructors` is not used (LK
init hooks use lk_init_level) so we disable the CFI check
unconditionally for simplicity.

Bug: 230134581
Change-Id: I78c00bee3941f5584f6db57596e868ba632a6404