lk: Add kernel BTI support.
Add support for ARM-A v8.5 Branch Target Identification.  This is a mandatory Aarch64 feature that checks the BR and BLR instructions land on a BTI instruction.  If this happens in a 'guarded' memory page (a new GP bit in the PTE is set), an exception is raised and execution can be halted (panic).  This resists Jump Oriented Programming (JOP) attacks by reducing available code 'gadgets'

Bug: 260397427
Test: build.py qemu-generic-arm64-test-debug
Test: build.py qemu-generic-arm32-test-debug
Test: build.py qemu-generic-arm64-test-debug, Qemu 7 with cpu=max (BTI
supported)

Change-Id: I682cb40be61b4819caa199e65338beeb452da766