include/shared: Make INCBIN sections allocatable
When the .incbin assembler directive doesn't recognize a
section by its name, it creates a new section without any
flags. This means that it creates sections that are not
allocatable so the sections don't end up in the final
executables which leads to linker errors if something
references the embedded binary. This commit makes all
sections created by INCBIN allocatable so the linker will
place them in the executables.

Bug: 231151995
Change-Id: Iac88cb0c6581167209b04ba4b1640887f8eef47b