engine.mk: LOCAL_MAKEFILE needs to use lastword
To get the current makefile, you need to get the last value of
MAKEFILE_LIST before any include statements in the makefile. This
one place was getting the whole list of makefiles. If you have an
environment with one or more convenience makefiles in the "MAKEFILES"
env variable, then this breaks.

Bug: 279109533
Test: build with MAKEFILES set to /tmp/foo.mk
Change-Id: I0feadee779c71e7ac009fdce3db50229fb8ddf0d